(fix): projects were not showing before
